Add Formatter class to follow the Single Responsibility Principle for TextUI
