+++ src/seedu/addressbook/ui/Formatter.java
@@ -1,0 +1,50 @@
+package seedu.addressbook.ui;
+
+import java.util.List;
+
+public class Formatter {
+
+    /** Offset required to convert between 1-indexing and 0-indexing.  */
+    public static final int DISPLAYED_INDEX_OFFSET = 1;
+    
+    /** A platform independent line separator. */
+    protected static final String LS = System.lineSeparator();
+    
+    /** A decorative prefix added to the beginning of lines printed by AddressBook */
+    protected static final String LINE_PREFIX = "|| ";
+
+    protected static final String DIVIDER = "===================================================";
+
+    /** Format of indexed list item */
+    private static final String MESSAGE_INDEXED_LIST_ITEM = "\t%1$d. %2$s";
+    
+    /** Format message(s) going to be shown to the user */
+    protected static String formatMessageShownToUser(String... message) {
+        String output = "";
+        for (String m : message) {
+            output += LINE_PREFIX + m.replace("\n", LS + LINE_PREFIX) + "\n";
+        }
+        return output;
+    }
+    
+    /** Formats a list of strings as a viewable indexed list. */
+    protected static String getIndexedListForViewing(List<String> listItems) {
+        final StringBuilder formatted = new StringBuilder();
+        int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
+        for (String listItem : listItems) {
+            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
+            displayIndex++;
+        }
+        return formatted.toString();
+    }
+    
+    /**
+     * Formats a string as a viewable indexed list item.
+     *
+     * @param visibleIndex visible index for this listing
+     */
+    private static String getIndexedListItem(int visibleIndex, String listItem) {
+        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
+    }
+
+}
